Make file locking logic more generic
This removes the references to a "history" file from the file locking logic.
The locking is not history specific, and useful for other applications, so
prepare it to be extracted to a place where it can be used by
non-history-related code.

The only required change in the logic is to pass the entire file path to the
constructor, instead of using `history_filename` in the constructor.