Combine file locking implementations
The intention is to reduce the number of separate locking implementations.
This provides a shared locking implementation for the history and universal
variable files.

The most important change is that the lock+rewrite operations can now be handled
by the `rewrite_via_temporary_file` function.
This function sets up the right context for calling the `rewrite` function it
gets as its argument.
First, a temporary file is created and set up in such a way that it gets
deleted when the function returns.
Then, we have one code path for working file locks, and one fallback.
If locking works:
  - Lock the file.
  - Call the `rewrite` function. Fail on error.
  - Update the metadata of the temporary file.
  - Rename to replace the old version of the file with the temporary file.
  - Obtain the `FileId` for the path which now stores the new file.
  - Return this id (and any data returned by `rewrite`).
  - Upon returning the lock is released.
If locking does not work, it gets more complicated. Instead of preventing races,
we try to detect and mitigate them. When we detect concurrent modification, we
retry, up to a fixed number of attempts.
In each attempt, the following is done:
  - Truncate the temporary file to length 0 to avoid data left over from
    previous attempts.
  - Get the `FileId` for the path.
  - Open (and possibly create) the file.
  - Abort attempt if the `FileId` for the opened file indicates that it changed.
  - Call the `rewrite` function. Abort attempt on error.
  - Update the metadata of the temporary file.
  - Check the `FileId` for the path again, abort on change.
  - Now the critical section where we do not detect races begins.
  - Rename to replace the old version of the file with the temporary file.
  - Get the `FileId` for the path again. This id is returned, alongside any data
    returned by `rewrite`. Note that this id might differ from the one we could
    obtain from the temporary file due to the rename, so we cannot use this to
    detect a conflict.

At the moment, locking is not disabled if previous attempts took "too long",
only on failure and in advance if a remote file system is detected.

Locking via `O_EXLOCK` which was used for the universal variable file is no
longer supported. This version of locking locks the file we operate on directly,
instead of using the lock file. According to the man pages of {Open,Free,Net}BSD
and macos, these locks have flock semantics. So if flock is available, we can
just use it, and if it is not, the `O_EXLOCK` flag does not help.

The code for saving the universal variable file required fairly extensive
modification to be compatible with the new locking logic.
This also includes a behavioral change: To ensure that we can safely retry when
locking is unavailable, the `EnvUniversal` struct is only updated if saving
succeeds.