Remove `EnvUniversal::ok_to_save`
The only place setting this aside from initialization to `true` is in
`load_from_file`, which sets it to `false` when an unknown file format is
detected.

Due to the better error handling introduced earlier, we no longer need
`ok_to_save` to decide whether saving should be attempted.

The `test_universal_ok_to_save` test used a getter for this field for some
assertions. Because the error returned by `load_from_file` is not propagated to
public functions accessible to the test, this test loses some power, as the
return values of the functions it uses to modify the state cannot be used to
uniquely identify the "unknown format" error.
To resolve this, it would be necessary to propagate errors further up, or to
make the `load_from_file` function accessible to the test.

Now `load_from_file` can almost use non-mutable `self`. The only holdout is
`self.last_read_file_id`, which will be replaced along with a unified locking
and fallback implementation.