Use proc macro for gettext extraction
This is a replacement for the `cargo-expand`-based extraction.

When building with the `GETTEXT_EXTRACTION_DIR` environment variable set, the
`gettext_extract` proc macro will write the messages passed to the `wgettext!`
macro to a file in the directory specified by the variable.

The macro can handle string literals, which are written to the `literals` file
enclosed in double quotes and with a trailing newline.
Passing the name of string constants is also supported, but the proc macro is
unable to obtain the corresponding literal, so instead it writes the name of the
constant into the `consts` file. Actually extracting the string is then done in
`build_tools/fish_xgettext.fish`, which uses the identifiers from `consts` to
search the source files for the corresponding definitions.