Use proc macro for gettext extraction
This is a replacement for the `cargo-expand`-based extraction.

When building with the `FISH_GETTEXT_EXTRACTION_DIR` environment variable set, the
`gettext_extract` proc macro will write the messages passed to the `wgettext!`
macro to a file in the directory specified by the variable.

The macro can handle string literals, which are written to the `literals` file
enclosed in double quotes and with a trailing newline.
Passing the name of string constants is also supported, but the proc macro is
unable to obtain the corresponding literal, so instead it writes the name of the
constant into the `consts` file. Actually extracting the string is then done in
`build_tools/fish_xgettext.fish`, which uses the identifiers from `consts` to
search the source files for the corresponding definitions.

Updates to the po files:
- This is the result of running the `update_translations.fish` script using the
  new proc_macro extraction. It finds 10 additional messages compared to the
  cargo-expand based approach.
- Messages IDs corresponding to paths are removed. The do not have translations
  in any language and translating paths would not make sense. I have not
  investigated how they made it into the po files in the first place.
- The only msgid which is removed that had translations is "Could not show help
  message". This literal corresponds to the `HELP_ERR` constant, which is only
  used in `eprintf!` calls, so they are never translated.

Remove docs about installing `cargo-expand`
These are no longer needed due to the switch to our extraction macro.