Regenerate po files
This is the result of running the `update_translations.fish` script using the
new proc_macro extraction. It finds 10 additional messages compared to the
cargo-expand based approach.

Messages IDs corresponding to paths are removed. The do not have translations in
any language and translating paths would not make sense. I have not investigated
how they made it into the po files in the first place.

The only msgid which is removed that had translations is "Could not show help
message". This literal corresponds to the `HELP_ERR` constant, which is only
used in `eprintf!` calls, so they are never translated.