Avoid leaking gettext message ids unnecessarily
Now that the gettext cache is locale-aware, we might leak a non-static message
id multiple times if it is used in multiple locales. To avoid this, we add a set
for caching the static references created by leaking the key the first time,
which is checked before attempting to leak a key, so we avoid leaking the key a
second time.