Add gettext localization test
This test requires certain locales, so it might not run in most circumstances.
In CI, we can generate the locales we need, so at least there the test should
run without issues.

In order to find the mo files for localization, the install step is added to the
CI jobs.