Fix gettext for CMake builds
The MO files are automatically generated and should therefore not be placed in a
source directory. Instead, use `$FISH_BUILD_DIR/fish-locale`.
For a cargo-only build, this is `target/fish-locale`.
For a CMake build, it is `build/fish-locale` (assuming `build` is CMake's build
directory).

Logic in `src/env/config_paths.rs` inspects the path of fish's own binary when
running, and based on it decides where to look for the MO files.
If the binary is located in `$FISH_BUILD_DIR` (including sub-directories)
the files in `$FISH_BUILD_DIR/fish-locale` will be used.

Otherwise, the value of `$LOCALEDIR` at compile time will be used.
This is defined in our CMake builds.
Note that installing to a different prefix will break this unless the prefix was
already defined for building, e.g. via:
`cmake .. -DCMAKE_INSTALL_PREFIX=/nonstandard/prefix && make install`

Cargo does not support installing asset files aside from the binaries
themselves. For now, we do not embed MO files in the binary.
One reason is that for consistency, we would first have to compile once to
extract an up-to-date version of all localizable strings from the sources,
then generate the MO files from the results, followed by another compilation to
embed the updated MO files. The second compilation is not required for debug
builds, since these actually access the files on the file system instead of
embedded versions.
Another, more significant reason for not embedding MO files is that the system's
gettext implementation needs actual files on the file system, meaning that fish
would have to write out the embedded content into files, which is a bit wasteful
and comes with the issue of finding a suitable place for it.
Alternatively, we could bundle a gettext implementation in Rust which works with
the embedded files, eliminating the runtime dependency on a system-provided
gettext.