Revert "Restore terminal state on SIGTERM again"
This reverts commit 1d6fa258f670f2e9664f123fddc9d42e17b3c921.

This reintroduces commit 941701da3d8, which was then reverted in
941701da3d8; this commit reverts the revert to reintroduce 941701da3d8.

The reason is that the existing logic in terminal_protocols_disable_ifn does a
bunch of stuff for which nobody has thought about its signal safety, such as
accessing the reader stack (clearly not async signal safe).
Even functions which happen to be safe now may become unsafe in the future.

This is just the nature of signal handling code. We must ensure that only
async-signal safe syscalls are run, and only functions which are themselves
async-signal safe, which we (try) to designate with the "safe_" prefix.