Add docs for distros supporting native builds
Document on which distros we aim to support building fish using only
dependencies packaged for the distro.
The intention is providing clarity about which dependency versions can be used
safely.

For now, this only contains Rust versions, but it might also be useful for other
dependencies, like Python versions. Once we clearly indicate distro version
support, it becomes significantly easier to figure out if a dependency version
can be changed.

Inspired in part by discussion
https://github.com/fish-shell/fish-shell/discussions/11584