Use ll_CC.po for all PO files
The format without a country code is not supported on all platforms (e.g.
macOS).

Making the variant explicit also eliminates ambiguity for translators and users.

One downside of this approach is that the default fallback logic when a variant
is unavailable is to use English, instead of trying other variants of the same
language. This can be changes via the `LANGUAGE` variable.
https://www.gnu.org/software/gettext/manual/html_node/The-LANGUAGE-variable.html