Simplify gettext implementation
Caching translations is problematic, because the translation of a message might
change at runtime based on several environment variables, in particular,
`LANGUAGE`, `LC_ALL`, `LC_MESSAGES`, and `LANG`.
The logic for mapping the values of these variables to a list defining in which
order languages should by tried for translation is non-trivial and best left to
the actual gettext implementation.

These changes simplify the logic in fish a bit and make language changes at
runtime work correctly, at the cost of always needing an allocation when gettext
is called. I think this is an acceptable tradeoff, which might become
unnecessary if we decide to bundle a gettext implementation instead of relying
on the one provided by the system.

We also call the system's gettext for each localization request in fish now.
This should not introduce much overhead, since gettext has internal caching,
which gets invalidated when `_nl_msg_cat_cntr` is changed.
We do this in `src/env_dispatch.rs` when a locale-related environment variable
is changed.
https://www.gnu.org/software/gettext/manual/html_node/gettext-grok.html