Added ability to specify that argparse should delete an option from both $argv and $argv_opts.
The intention is that if you want to parse some of your options verbatim to another command, but you want to modfy other options
(e.g. change their value, convert them to other options, or delete them entirely), you mark the options you want to modify with an &,
and argparse will not add them to argv_opts. You can then call the other command with argv_opts together with any new/modified options,
ensuring that the other command doesn't set the pre-modified options. As with other known options, & options will be removed from $argv,
and have their $_flag_ variables set.

The `&` goes at the end of the option spec, or if the option spec contains a validation script, immediately before the `!`.
There is also now a -d/--delete flag to fish_opt that will generate such an option spec.

See the changes in doc_src/cmds/argparse.rst for more details and an example use case.