Added an -L/--strict-longopts option to argparse.
This flag disables a very surprising and confusing feature I found in the code of wgetopt.rs:
the ability to abbreviate the names of long options and the ability to parse long options with a single "-".
This commit fixes #7341, but unlike pull request #11220, it does so in a backwards compatible way:
one must use the new -L/--strict-longotps flag to disable the old legacy behaviour.

Unlike pull request #11220 however, this flag only applies to ``argparse``, and not to any builtins
used by fish.

Note that forcing the flag -L/--strict-longotps on (i.e. in  src/wgetopt.rs, replacing both uses of `self.strict_long_opts` with
`true`), does not cause any of the current test cases to fail. However, third-party fish scripts may be depending
on the current behaviour.