Added a -m/--move-unknown option to argparse.
--move-unknown is like --ignore-unknown, but unknown options are instead moved from $argv to $argv_opts, just like known ones.
This allows unambiguously parsing non-option arguments to other commands. For example if $argv contains `--opt -- --file`, and we execute
`argparse --move-unknown -- $argv`, we can then call `cmd $argv_opts -- --another-file $argv`, which will correctly interpret `--opt` as an option,
but `--file` and `--some-file` as an argument.
This makes `--move-unknown` a better alternative to `--ignore-unknown`, however the latter is kept for backwards compatibility.