Added an -a/--unknown-arguments option to argparse, altering how arguments to unknown options are parsed.
The new -a/--unknown-arguments option takes either 'optional', 'required', or 'none', indicating how many arguments unknown options are assumed
to take. The default is optional, the same behaviour as before this commit, despite most options in practice taking not taking any arguments.
Using --unknown-arguments=required and --unknown-arguments=none (but not --unknown-arguments=optional) can give you parse errors if, respectively,
an unknown option has no argument (because it the option is at the end of the argument list), or is given an argument (with the `--flag=<value> syntax).
See doc_src/cmds/argparse.rst for more details (specifically, the descritpion of the --unknown-arguments flag and the example at the end
of the examples section).