Modified argparse to support one character long only options.
This fixes an issue noticed in the previous commit (the made the -s/--short option optional to fish_opt): it was impossible to define a single character
long flag, unless you also provided a single-character short flag equivalent.

This commit works by allowing an option spec to start with a '/', treating the subsequent alpha-numeric characters as a long flag name.

In detail, consider the following:
- s defines a -s short flag
- ss defines an --ss long flag
- /ss (new) also defines a --ss long flag
- s/s defines a -s short flag and an --s long flag
- s-s defines a --s long flag (if there's already an -s short flag, you'd have to change the first s, e.g. S-s)
- /s (new) defines a --s long flag
- s/ is an error (a long flag name must follow the /)
Note that without using the --strict-longopts, a long flag --s can always be abbreviated as -s, provided that -s isn't defined as a separate short flag.

This 'issue' fixed by this commit is relatively trivial, however it does allow simplifying the documentation for fish_opt (since it no longer needs to mention the restriction).
In particular, this commit makes the --long-only flag to fish_opt completely unnecessary (but it is kept for backwards compatibility).